Time: 48 ms (83.06%), Space: 22.1 MB (96.34%) - LeetHub

+++ 2221-find-triangular-sum-of-an-array/2221-find-triangular-sum-of-an-array.cpp
@@ -20,7 +20,7 @@ public:
     int triangularSum(vector<int>& nums) {
         int n = nums.size();
         while (n > 1) {
-            for (int i = 0; i < n - 1; ++i) {
+            for (size_t i = 0; i < n - 1; ++i) {
                 nums[i] = (nums[i] + nums[i + 1]) % 10;
             }
             --n;
